// src/domain/horasExtras/types.ts
export interface HorasExtras {
    id: number;
    funcionario_id: number;
    solicitante_id: number;
    data_solicitacao: string;
    data_extra: string;
    horas_extras: string;
    observacao: string;
    autorizado_por: number;
    data_autorizacao: string;
    status_solicitacao_id: string;
}